+++ src/main/java/com/hseong/jungleproduct/controller/request/AddProductRequest.java
@@ -1,4 +1,4 @@
-package com.hseong.jungleproduct.controller;
+package com.hseong.jungleproduct.controller.request;
 
 public record AddProductRequest(Long productId, String name, int price) {
 }
