+++ src/main/java/com/hseong/jungleproduct/controller/request/BuyProductRequest.java
@@ -1,4 +1,4 @@
-package com.hseong.jungleproduct.controller;
+package com.hseong.jungleproduct.controller.request;
 
 public record BuyProductRequest(Long productId, int amount) {
 }
