+++ src/main/java/com/hseong/jungleproduct/controller/request/UpdateProductRequest.java
@@ -1,4 +1,4 @@
-package com.hseong.jungleproduct.controller;
+package com.hseong.jungleproduct.controller.request;
 
 public record UpdateProductRequest(int displayAmount, int storageAmount) {
 }
