+++ src/test/java/com/hseong/jungleproduct/controller/OrderControllerTest.java
@@ -9,6 +9,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.hseong.jungleproduct.base.BaseControllerTest;
+import com.hseong.jungleproduct.controller.request.BuyProductRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
