+++ src/test/java/com/hseong/jungleproduct/controller/ProductControllerTest.java
@@ -11,18 +11,15 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hseong.jungleproduct.base.BaseControllerTest;
+import com.hseong.jungleproduct.controller.request.AddProductRequest;
+import com.hseong.jungleproduct.controller.request.UpdateProductRequest;
 import com.hseong.jungleproduct.service.ProductDto;
-import com.hseong.jungleproduct.service.ProductService;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
 @WebMvcTest
